Fix header anchors
Non-markdown heading components accept `text` as a plain string, so casting it
to a list and taking the first element only gets the first letter.

For markdown heading components, `text` is passed as "children" and is a list
of str.

To make both work, remove the `convert_to_str` kwarg, and dereference the first
element inside the `_xd` memo components which are used by markdown.